Move dir to go
