--- go.mod
@@ -1,3 +1,0 @@
-module github.com/daido1976/monkey
-
-go 1.16
