+++ go/go.mod
@@ -1,0 +1,3 @@
+module github.com/daido1976/monkey/go
+
+go 1.17
